d/*.symbols: Update symbols files for 4.1
Addition of new functions are easy to handle, but adding new fields to
the end of structures can't be handled directly by the deb-symbols
system. This is worked around by bumping the version for a few
functions which must be called by the calling program to be able to
use these structures. It is possible to defeat this if: the structure
crosses a (non-ffmpeg) shared library boundary, the consumer of said
library does not call one of the bumped functions, and the consumer
references a new field. Hopefully these cases are rare enough that
they can be fixed manually in the consuming packages (if there are any
cases at all).

libavcodec:
+ av_bsf_flush
+ av_mediacodec_render_buffer_at_time

libavformat:
+ struct AVFormatContext.skip_estimate_duration_from_pts
  - avformat_alloc_context
  - avformat_alloc_output_context
  - avformat_open_input
+ struct AVProgram.pmt_version
  - av_find_program_from_stream
  - av_new_program
  - Note you can also obtain an AVProgram if you have an
    AVFormatContext object, but the version bumps above should cover
    this.

libavutil:
+ struct AVEncryptionInitInfo.next.
  - av_encryption_init_info_alloc
  - av_encryption_init_info_get_side_data
+ av_read_image_line2
+ av_thread_message_queue_nb_elems
+ av_write_image_line2