Introduce new binary package ffplay (Closes: #923494)
The ffplay utility is an SDL application, which these days comes
with a lot of sizable dependencies. In a headless installation that
is used for transcoding and streaming, such dependencies, like on
X11, wayland, etc. may not be desirable.

This also disables the sdl2 output device, which according to Carl-Eugen
from FFmpeg upstream is mostly used for debugging.

With this patch, the ffmpeg binary package has a depends line like this:

 Package: ffmpeg
 Version: 7:4.1.1-2
 Architecture: amd64
 Maintainer: Debian Multimedia Maintainers <debian-multimedia@lists.debian.org>
 Installed-Size: 1808
 Depends: libavcodec58 (= 7:4.1.1-2), libavdevice58 (= 7:4.1.1-2), libavfilter7 (= 7:4.1.1-2), libavformat58 (= 7:4.1.1-2), libavresample4 (= 7:4.1.1-2), libavutil56 (= 7:4.1.1-2), libc6 (>= 2.14), libpostproc55 (= 7:4.1.1-2), libswresample3 (= 7:4.1.1-2), libswscale5 (= 7:4.1.1-2)
 Suggests: ffmpeg-doc
 Breaks: libav-tools (<< 6:12~~), qt-faststart (<< 7:2.7.1-3~), winff (<< 1.5.5-5~)
 Replaces: libav-tools (<< 6:12~~), qt-faststart (<< 7:2.7.1-3~)
 Section: video

Note that there is a dependency on libavdevice58, but not on SDL.

The 'ffplay' binary package has a depends line that looks like this:

 Package: ffplay
 Source: ffmpeg
 Version: 7:4.1.1-2
 Architecture: amd64
 Maintainer: Debian Multimedia Maintainers <debian-multimedia@lists.debian.org>
 Installed-Size: 226
 Depends: libavcodec58 (= 7:4.1.1-2), libavdevice58 (= 7:4.1.1-2), libavfilter7 (= 7:4.1.1-2), libavformat58 (= 7:4.1.1-2), libavresample4 (= 7:4.1.1-2), libavutil56 (= 7:4.1.1-2), libc6 (>= 2.14), libpostproc55 (= 7:4.1.1-2), libsdl2-2.0-0 (>= 2.0.9), libswresample3 (= 7:4.1.1-2), libswscale5 (= 7:4.1.1-2), ffmpeg
 Suggests: ffmpeg-doc
 Breaks: ffmpeg (<< 7:4.1.1-2~), libav-tools (<< 6:12~~), qt-faststart (<< 7:2.7.1-3~), winff (<< 1.5.5-5~)
 Replaces: ffmpeg (<< 7:4.1.1-2~), libav-tools (<< 6:12~~), qt-faststart (<< 7:2.7.1-3~)
 Section: video

Note that this includes both libavdevice58 as well as libsdl2-2.