gitattributes: Disable CRLF normalisation in fate test
FFmpeg contains one file with CRLF line endings which dgit doesn't
like (because the checked out file != the version in git). Disable
CRLF normalisation on that file so it's committed as CRLF instead.

Note that dgit automatically overrides this setting if you run "dgit
setup-gitattributes" or just clone the repository using dgit. This
change is mainly to prevent non-dgit users from importing a new
upstream with line endings which don't work with dgit.